Port scenario 2 changes to new file
